unit: add waiting jobs to run queue in unit_coldplug
When we have job installed and added to run queue for service which is still in
dead state and systemd initiates reload then after reload we never add
deserialized job to the run queue again. This is caused by check in
service_coldplug() where we check if deserialized state is something else than
dead state, which is not the case thus we never call service_set_state() and
finally unit_notify() where we would have added job to the run queue.