udev: net_id - correctly name netdevs based on dev_port when set
Upstream, dev_id was replaced by dev_port, and the same happened for some kernel
drivers. This logic is not in the RHEL7 kernel, except for one new driver which
uses dev_port, but never used dev_id in the past.

To give proper names to these devices, fall back to using dev_port when dev_id
is not set. This does not affect any existing drivers.

(rhel only)

Resolves: #1155996

Conflicts:
	src/udev/udev-builtin-net_id.c

Conflicts:
	src/udev/udev-builtin-net_id.c